Add realtime BPM tags and large cleanup
This change adds the realtime bpm tags that will update output files
anytime the bpm changes. This also includes a large amount of fixes,
enhancements, and code cleanup.

This improves the injector to inject before Rekordbox launches.

This includes the new hooking code which uses capstone to measure the
length of opcodes to relocate.

This includes several small bugfixes which were previously unnoticed